+++ rcn_py/topic_modeling.py
@@ -1,9 +1,11 @@
 import nltk
 from nltk.corpus import stopwords
-from nltk.tokenize import word_tokenize
 from nltk.stem import WordNetLemmatizer
 from sklearn.feature_extraction.text import TfidfVectorizer
 from sklearn.decomposition import LatentDirichletAllocation
+import gensim
+from gensim import corpora
+from rcn_py import openalex
 
 def build_corpus(publications, num_topics, max_ngram_length=2):
     # Download necessary NLTK resources
@@ -53,3 +55,38 @@ def build_corpus(publications, num_topics, max_ngram_length=2):
         topics.append(top_words)
         
     return groups,topics
+
+
+def openalex_build_corpus(works):
+    keywords_list = []
+    for w in works:
+        if (w['doi']):
+            keywords = openalex.publication_keywords(w['doi'])
+            keywords_list.append(keywords)
+        else:
+            continue
+
+    dictionary = corpora.Dictionary(keywords_list)
+    corpus = [dictionary.doc2bow(keywords) for keywords in keywords_list]
+
+    # Train LDA Model
+    num_topics = 10  # Specify the number of topics
+    lda_model = gensim.models.LdaModel(corpus, num_topics=num_topics, id2word=dictionary, passes=10)
+
+    # Get Topics and Assign Topic Names
+    topics = lda_model.print_topics(num_words=5)  # Specify the number of words per topic
+
+    # Print and Save Topic Names
+    topic_names = {}
+    for i, publication in enumerate(works):
+        publication_topics = lda_model.get_document_topics(corpus[i])
+        top_topic = max(publication_topics, key=lambda x: x[1])[0]  # Get the topic with the highest weight for the publication
+        top_keywords = [word.split("*")[1].strip().replace('"', '') for word in topics[top_topic][1].split("+")]
+        # topic_names[i] = ", ".join(top_keywords)
+        topic_names[i] = {"topic_name": ", ".join(top_keywords), "group_id": top_topic}
+        publication["topic_name"] = top_keywords
+        publication["group_id"] = top_topic
+
+    # Return topic names for each publication
+    return topic_names, works
+
