+++ static/D3_js/networks/side_bar.js
