+++ static/D3_js/zoom_drag/canvas_move.js
